feat(web): rename file
